+++ Events.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace VAMP;
+
+public class Events
+{
+    public static Action OnCoreLoaded;
+}