+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 `latest`
+- Updated for 1.1 Hotfix 3
+- Added ChatUtil.SystemSendAdmins
+- PlayerService now uses Player model
+- Old PlayerService methods marked deprecated
+- Added static Events class with OnCoreLoaded (more coming next update)
+- Added CleanItemName to ItemUtil
+- Updated Documentation for PlayerService and ChatUtil
+
+`0.2.2`
 - Added ChatUtil.SystemSendAllExcept
 - Added IsAdminCapable to Player
 - Added new ways to convert to Player model from Character Entity and User